+++ instrumentation/github.com/gorilla/mux/otelmux/internal/semconvutil/httpconv_test.go
@@ -68,6 +68,29 @@ func TestHTTPSClientRequest(t *testing.T) {
 	)
 }
 
+func TestHTTPSClientRequestMetrics(t *testing.T) {
+	req := &http.Request{
+		Method: http.MethodGet,
+		URL: &url.URL{
+			Scheme: "https",
+			Host:   "127.0.0.1:443",
+			Path:   "/resource",
+		},
+		Proto:      "HTTP/1.0",
+		ProtoMajor: 1,
+		ProtoMinor: 0,
+	}
+
+	assert.ElementsMatch(
+		t,
+		[]attribute.KeyValue{
+			attribute.String("http.method", "GET"),
+			attribute.String("net.peer.name", "127.0.0.1"),
+		},
+		HTTPClientRequestMetrics(req),
+	)
+}
+
 func TestHTTPClientRequest(t *testing.T) {
 	const (
 		user  = "alice"
@@ -102,6 +125,40 @@ func TestHTTPClientRequest(t *testing.T) {
 			attribute.Int("http.request_content_length", n),
 		},
 		HTTPClientRequest(req),
+	)
+}
+
+func TestHTTPClientRequestMetrics(t *testing.T) {
+	const (
+		user  = "alice"
+		n     = 128
+		agent = "Go-http-client/1.1"
+	)
+	req := &http.Request{
+		Method: http.MethodGet,
+		URL: &url.URL{
+			Scheme: "http",
+			Host:   "127.0.0.1:8080",
+			Path:   "/resource",
+		},
+		Proto:      "HTTP/1.0",
+		ProtoMajor: 1,
+		ProtoMinor: 0,
+		Header: http.Header{
+			"User-Agent": []string{agent},
+		},
+		ContentLength: n,
+	}
+	req.SetBasicAuth(user, "pswrd")
+
+	assert.ElementsMatch(
+		t,
+		[]attribute.KeyValue{
+			attribute.String("http.method", "GET"),
+			attribute.String("net.peer.name", "127.0.0.1"),
+			attribute.Int("net.peer.port", 8080),
+		},
+		HTTPClientRequestMetrics(req),
 	)
 }